[refactor] HttpRequestParser 클래스 분리
HttpRequestParser 가 너무 많은 메서드를 가지고 있어서 가독성이 매우 떨어졌습니다.
적절한 범위로 메서드들을 별도의 클래스로 분리하여 가독성을 높였습니다.
